% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{feature_select}
\alias{feature_select}
\title{feature_select}
\usage{
feature_select(pv_data_in, recode_bills = T, drop_lowestbills = F)
}
\arguments{
\item{pv_data_in}{pv_survey_oo or pv_survey}

\item{recode_bills}{if true an annual bill feature q_ab is introduced with values}

\item{drop_lowestbills}{if true q15 is dropped}
}
\value{
reduced survey dataframe
}
\description{
removes selected features from pv survey dataset before calibration
if recode_bills=T an annual bill (q_ab) feature is introduced and min-max bills are dropped
}
