% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{get_shap_scores}
\alias{get_shap_scores}
\title{get_shap_scores}
\usage{
get_shap_scores(pv_data_in, bst)
}
\arguments{
\item{pv_data_in}{the pv_survey dataset}

\item{bst}{xgboost model}
}
\value{
dataframe long format
}
\description{
get_shap_scores
}
